# src/typescript/coreapis-sdk/openapi.yaml
openapi: 3.0.1
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /v1/users:
    get:
      tags:
      - Users
      summary: List Users
      operationId: list_users
      security:
      - SecurityChecker: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 25
          description: Items per page
          default: 25
          title: Limit
        description: Items per page
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/me/info:
    get:
      tags:
      - Users
      summary: Current User Info
      operationId: user_info
      responses:
        '200':
          description: User Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - SecurityChecker: []
  /v1/users/registration:
    post:
      tags:
      - Users
      summary: User Registration
      operationId: register_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreatePayload'
        required: true
      responses:
        '201':
          description: Operation status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/me/update:
    put:
      tags:
      - Users
      summary: Update current user
      operationId: update_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdatePayload'
        required: true
      responses:
        '200':
          description: User Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - SecurityChecker: []
  /v1/users/me/avatar:
    post:
      tags:
      - Users
      summary: User Avatar Upload
      operationId: user_avatar_upload
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_user_avatar_upload'
        required: true
      responses:
        '200':
          description: User Avatar Url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileInfo'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - SecurityChecker: []
...
components:
  schemas:
    AccountTypeEnum:
      type: string
      enum:
      - Admin
      - Agronomist
      - Standard
      title: AccountTypeEnum
...
    Body_user_avatar_upload:
      properties:
        avatar:
          type: string
          format: binary
          title: Avatar
          description: avatar as UploadFile
      type: object
      required:
      - avatar
      title: Body_user_avatar_upload
    Contacts:
      properties:
        email:
          type: string
          title: Email
        phone:
          type: string
          title: Phone
      type: object
      required:
      - email
      - phone
      title: Contacts
...
    ErrorResponse:
      properties:
        status:
          type: integer
          title: Status
        detail:
          anyOf:
          - type: string
          - type: object
          title: Detail
      type: object
      required:
      - status
      - detail
      title: ErrorResponse
...
    FileInfo:
      properties:
        category:
          type: string
          title: Category
        name:
          type: string
          title: Name
      type: object
      required:
      - category
      - name
      title: FileInfo
...
    OrganizationCreatePayload:
      properties:
        name:
          type: string
          title: Name
        piva:
          type: string
          title: Piva
        contacts:
          $ref: '#/components/schemas/Contacts'
        questionnaire:
          type: object
          additionalProperties: true
          title: Questionnaire
      type: object
      required:
      - name
      - piva
      - contacts
      - questionnaire
      title: OrganizationCreatePayload
...
    PaginatedResponse:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        limit:
          type: integer
          title: Limit
        data:
          items: {}
          type: array
          title: Data
      type: object
      required:
      - total
      - page
      - limit
      - data
      title: PaginatedResponse
...
    StatusResponse:
      properties:
        status:
          type: integer
          title: Status
        message:
          type: string
          title: Message
      type: object
      required:
      - status
      - message
      title: StatusResponse
    User:
      properties:
        id:
          type: string
          title: Id
        firstName:
          type: string
          title: Firstname
        lastName:
          type: string
          title: Lastname
        email:
          type: string
          title: Email
        emailVerified:
          type: boolean
          title: Emailverified
        enabled:
          type: boolean
          title: Enabled
        accountType:
          $ref: '#/components/schemas/AccountTypeEnum'
        phone:
          type: string
          title: Phone
        piva:
          type: string
          title: Piva
        avatar:
          type: string
          title: Avatar
        organizations:
          items:
            $ref: '#/components/schemas/UserOrgananizationMembership'
          type: array
          title: Organizations
          default: []
        creationTime:
          type: integer
          title: Creationtime
      type: object
      required:
      - id
      - firstName
      - lastName
      - email
      - emailVerified
      - enabled
      - accountType
      - phone
      - piva
      - avatar
      - creationTime
      title: User
    UserCreatePayload:
      properties:
        firstName:
          type: string
          title: Firstname
        lastName:
          type: string
          title: Lastname
        email:
          type: string
          title: Email
        accountType:
          $ref: '#/components/schemas/AccountTypeEnum'
        phone:
          type: string
          title: Phone
        piva:
          title: Piva
          type: string
          nullable: true
        organization:
          $ref: '#/components/schemas/OrganizationCreatePayload'
          nullable: true
      type: object
      required:
      - firstName
      - lastName
      - email
      - accountType
      - phone
      title: UserCreatePayload
    UserOrgananizationMembership:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        roles:
          items:
            type: string
          type: array
          title: Roles
      type: object
      required:
      - id
      - name
      - roles
      title: UserOrgananizationMembership
    UserUpdatePayload:
      properties:
        firstName:
          title: Firstname
          type: string
          nullable: true
        lastName:
          title: Lastname
          type: string
          nullable: true
        phone:
          title: Phone
          type: string
          nullable: true
      type: object
      title: UserUpdatePayload
  securitySchemes:
    SecurityChecker:
      type: http
      scheme: bearer